Cache memory silly limits

+++ src/libfrozen/data/memory/memory_t.h
@@ -16,9 +16,9 @@ typedef enum memory_t_alloc {
 
 typedef struct memory_t {
 	void           *ptr;
-	size_t          ptr_size;
-	size_t          exact_size;
-	size_t          page_size;
+	uintmax_t       ptr_size;
+	uintmax_t       exact_size;
+	uintmax_t       page_size;
 	
 	int             zero;
 	memory_t_type   type;
@@ -27,13 +27,13 @@ typedef struct memory_t {
 
 API memory_t_type  memory_string_to_type(char *string);
 API memory_t_alloc memory_string_to_alloc(char *string);
-API void     memory_new(memory_t *memory, memory_t_type type, memory_t_alloc alloc, size_t page_size, int zero); 
+API void     memory_new(memory_t *memory, memory_t_type type, memory_t_alloc alloc, uintmax_t page_size, int zero); 
 API void     memory_free(memory_t *memory); 
-API ssize_t  memory_size(memory_t *memory, size_t *size); 
-API ssize_t  memory_resize(memory_t *memory, size_t new_size); 
-API ssize_t  memory_grow(memory_t *memory, size_t size, off_t *pointer); 
-API ssize_t  memory_shrink(memory_t *memory, size_t size); 
-API ssize_t  memory_translate(memory_t *memory, off_t pointer, size_t size, void **pointer_out, size_t *size_out); 
+API ssize_t  memory_size(memory_t *memory, uintmax_t *size); 
+API ssize_t  memory_resize(memory_t *memory, uintmax_t new_size); 
+API ssize_t  memory_grow(memory_t *memory, uintmax_t size, off_t *pointer); 
+API ssize_t  memory_shrink(memory_t *memory, uintmax_t size); 
+API ssize_t  memory_translate(memory_t *memory, off_t pointer, uintmax_t size, void **pointer_out, uintmax_t *size_out); 
 
 extern data_proto_t memory_t_proto;
